Move more interfaces.
