+++ scheduler/src/main/scala/mesosphere/marathon/core/instance/update/InstanceUpdateEffect.scala
@@ -1,0 +1,5 @@
+package mesosphere.marathon
+package core.instance.update
+
+/** The state change effect after applying a state change operation to the instance */
+trait InstanceUpdateEffect extends Product with Serializable
